requirements.txt
GDAL==3.3.2
keras==2.6.0
tensorflow_gpu==2.6.0
matplotlib==3.4.3
numpy==1.21.2
opencv_python==4.5.3.56
osgeo==0.0.1
Pillow==8.3.2
tensorflow==2.6.0
